Move some classes to gtc/definitions.py
