+++ src/gt4py/backend/module_generator.py
@@ -24,8 +24,9 @@ import jinja2
 import numpy
 
 from gt4py import utils as gt_utils
-from gt4py.definitions import AccessKind, Boundary, DomainInfo, FieldInfo, ParameterInfo
+from gt4py.definitions import AccessKind, DomainInfo, FieldInfo, ParameterInfo
 from gtc import gtir, gtir_to_oir
+from gtc.definitions import Boundary
 from gtc.passes.gtir_k_boundary import compute_k_boundary, compute_min_k_size
 from gtc.passes.gtir_pipeline import GtirPipeline
 from gtc.passes.oir_access_kinds import compute_access_kinds
