+++ src/gt4py/utils/__init__.py
@@ -24,17 +24,14 @@ from .base import (
     classmethod_to_function,
     classproperty,
     compose,
-    filter_mask,
     flatten,
     flatten_iter,
     get_member,
-    interpolate_mask,
     is_collection,
     is_identifier_name,
     is_iterable_of,
     is_mapping_of,
     jsonify,
-    listify,
     make_dir,
     make_local_dir,
     make_module_from_file,
