+++ src/gtc/dace/expansion.py
@@ -26,7 +26,6 @@ import gtc.oir as oir
 from eve import codegen
 from eve.codegen import FormatTemplate as as_fmt
 from eve.codegen import MakoTemplate as as_mako
-from gt4py.definitions import Extent
 from gtc.common import LoopOrder
 from gtc.dace.nodes import HorizontalExecutionLibraryNode, VerticalLoopLibraryNode
 from gtc.dace.utils import (
@@ -36,6 +35,7 @@ from gtc.dace.utils import (
     get_interval_range_str,
     get_tasklet_symbol,
 )
+from gtc.definitions import Extent
 from gtc.oir import Interval
 from gtc.passes.oir_optimizations.utils import AccessCollector
