+++ src/gtc/numpy/scalars_to_temps.py
@@ -20,8 +20,8 @@ from dataclasses import dataclass
 from typing import Dict
 
 from eve import NodeTranslator
-from gt4py.definitions import Extent
 from gtc import common, utils
+from gtc.definitions import Extent
 from gtc.numpy import npir
