+++ src/gtc/passes/horizontal_masks.py
@@ -16,8 +16,8 @@
 
 from typing import Optional, Tuple
 
-from gt4py.definitions import Extent
 from gtc import common
+from gtc.definitions import Extent
 
 
 def _overlap_along_axis(
