+++ src/gtc/passes/oir_access_kinds.py
@@ -19,8 +19,9 @@ import collections
 from typing import Any, Dict
 
 from eve.visitors import NodeVisitor
-from gt4py.definitions import AccessKind, Extent
+from gt4py.definitions import AccessKind
 from gtc import oir
+from gtc.definitions import Extent
 from gtc.passes.horizontal_masks import mask_overlap_with_extent
 from gtc.passes.oir_optimizations.utils import compute_horizontal_block_extents
