+++ tests/test_unittest/test_gtc/test_passes/test_oir_optimizations/test_utils.py
@@ -16,9 +16,9 @@
 
 import pytest
 
-from gt4py.definitions import Extent
 from gtc import common
 from gtc.common import DataType
+from gtc.definitions import Extent
 from gtc.passes.horizontal_masks import _overlap_along_axis, compute_relative_mask
 from gtc.passes.oir_optimizations.utils import (
     AccessCollector,
